+++ src/main/java/org/openlmis/auth/domain/BaseEntity.java
@@ -35,7 +35,8 @@ public abstract class BaseEntity implements Identifiable {
 
   @Id
   @GeneratedValue(generator = "uuid-gen")
-  @GenericGenerator(name = "uuid-gen", strategy = "uuid2")
+  @GenericGenerator(name = "uuid-gen",
+      strategy = "org.openlmis.auth.domain.ConditionalUuidGenerator")
   @Type(type = UUID_TYPE)
   @Getter
   @Setter