+++ src/main/java/org/openlmis/auth/service/notification/UserContactDetailsDto.java
@@ -5,19 +5,16 @@
  * This program is free software: you can redistribute it and/or modify it under the terms
  * of the GNU Affero General Public License as published by the Free Software Foundation, either
  * version 3 of the License, or (at your option) any later version.
- *  
+ *  
  * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
  * See the GNU Affero General Public License for more details. You should have received a copy of
  * the GNU Affero General Public License along with this program. If not, see
- * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
+ * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
  */
 
-package org.openlmis.auth.dto.referencedata;
+package org.openlmis.auth.service.notification;
 
-import com.google.common.collect.Sets;
-import java.util.Map;
-import java.util.Set;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -28,19 +25,13 @@ import lombok.ToString;
 
 @Getter
 @Setter
+@ToString
+@EqualsAndHashCode
 @NoArgsConstructor
 @AllArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
-public class UserMainDetailsDto extends BaseDto {
-  private String username;
-  private String firstName;
-  private String lastName;
-  private String jobTitle;
-  private String timezone;
-  private UUID homeFacilityId;
-  private boolean active;
-  private boolean loginRestricted;
-  private Map<String, String> extraData;
-  private Set<RoleAssignmentDto> roleAssignments = Sets.newHashSet();
+public final class UserContactDetailsDto {
+  private UUID referenceDataUserId;
+  private String phoneNumber;
+  private Boolean allowNotify;
+  private EmailDetailsDto emailDetails;
 }