+++ src/main/java/org/openlmis/auth/service/referencedata/UserReferenceDataService.java
@@ -15,8 +15,6 @@
 
 package org.openlmis.auth.service.referencedata;
 
-import java.util.Collections;
-import java.util.Map;
 import java.util.UUID;
 import lombok.Getter;
 import org.openlmis.auth.dto.ResultDto;
@@ -24,7 +22,6 @@ import org.openlmis.auth.dto.referencedata.UserMainDetailsDto;
 import org.openlmis.auth.service.BaseCommunicationService;
 import org.openlmis.auth.service.RequestParameters;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -44,29 +41,6 @@ public class UserReferenceDataService extends BaseCommunicationService<UserMainD
 
   protected Class<UserMainDetailsDto[]> getArrayResultClass() {
     return UserMainDetailsDto[].class;
-  }
-
-  /**
-   * This method retrieves a user with given email address.
-   *
-   * @param email the email address of user.
-   * @return UserMainDetailsDto containing user's data, or null if such user was not found.
-   */
-  public UserMainDetailsDto findUserByEmail(String email) {
-    Map<String, Object> payload = Collections.singletonMap("email", email);
-
-    Page<UserMainDetailsDto> users = getPage("search", Collections.emptyMap(), payload);
-    return users.getContent().isEmpty() ? null : users.getContent().get(0);
-  }
-
-  /**
-   * This method creates or updates a user.
-   *
-   * @param user the user to put.
-   * @return UserMainDetailsDto containing user's data.
-   */
-  public UserMainDetailsDto putUser(UserMainDetailsDto user) {
-    return put("", user);
   }
 
   /**