+++ src/test/java/org/openlmis/auth/DummyUserMainDetailsDto.java
@@ -33,10 +33,8 @@ public final class DummyUserMainDetailsDto extends UserMainDetailsDto {
    * Creates new instance of dummy user dto.
    */
   public DummyUserMainDetailsDto() {
-    super(
-        USERNAME, "Admin", "User", EMAIL, null, null, null, null, true,
-        false, false, true, null, getDummyRoleAssignments()
-    );
+    super(USERNAME, "Admin", "User", null, null, null, true, false, null,
+        getDummyRoleAssignments());
     setId(UUID.fromString(REFERENCE_ID));
   }