+++ src/test/java/org/openlmis/auth/service/notification/UserContactDetailsDtoTest.java
@@ -5,38 +5,34 @@
  * This program is free software: you can redistribute it and/or modify it under the terms
  * of the GNU Affero General Public License as published by the Free Software Foundation, either
  * version 3 of the License, or (at your option) any later version.
- *  
+ *  
  * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
  * See the GNU Affero General Public License for more details. You should have received a copy of
  * the GNU Affero General Public License along with this program. If not, see
- * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
+ * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
  */
 
-package org.openlmis.auth.dto.referencedata;
+package org.openlmis.auth.service.notification;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 import org.junit.Test;
-import org.openlmis.auth.DummyUserMainDetailsDto;
 import org.openlmis.auth.ToStringTestUtils;
 
-public class UserMainDetailsDtoTest {
-  private UserMainDetailsDto user = new DummyUserMainDetailsDto();
+public class UserContactDetailsDtoTest {
 
   @Test
   public void equalsContract() {
     EqualsVerifier
-        .forClass(UserMainDetailsDto.class)
-        .withRedefinedSuperclass()
-        .suppress(Warning.STRICT_INHERITANCE)
-        .suppress(Warning.NONFINAL_FIELDS) // dto can't contain final fields
+        .forClass(UserContactDetailsDto.class)
+        .suppress(Warning.NONFINAL_FIELDS)
         .verify();
   }
 
   @Test
   public void shouldImplementToString() {
-    ToStringTestUtils.verify(UserMainDetailsDto.class, user);
+    ToStringTestUtils.verify(UserContactDetailsDto.class, new UserContactDetailsDto());
   }
 
 }